get_next_line final tweaks

+++ get_next_line.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: linuxusr <linuxusr@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ceduard2 <ceduard2@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/21 22:35:44 by linuxusr          #+#    #+#             */
-/*   Updated: 2021/09/29 18:47:52 by linuxusr         ###   ########.fr       */
+/*   Updated: 2021/10/03 17:28:57 by ceduard2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 # endif
 
 size_t	ft_strlen(char *s);
-int		ft_search_nl(char *s);
+size_t	ft_search_nl(char *s);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_str_nl_cut(char *s);
 char	*ft_str_nl_rmd(char *left_str);
